backend/requirements.txt
# python 3.10
Django==5.1.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
django-cors-headers==4.7.0
# psycopg2-binary==2.9.10  # 如需连接PostgreSQL数据库
PyJWT>=1.5.2
sqlparse==0.5.0
asgiref>=3.8.1
openai>=1.55.3
python-dotenv==1.0.0
httpx==0.27.2
tushare>=1.2.89
pandas>=1.5.0
numpy>=1.21.0